Modifies the UIManager and UIElements to use pixel dimensions, adds some new (commented out) test code

+++ Mountainhome/MHUIElement.h
@@ -27,7 +27,7 @@ public:
     static VALUE AddChild(VALUE self, VALUE child);
 
 private:
-    float _width, _height;
+    int _width, _height;
     std::string _text;
 
 	MHUIManager *_manager;
